Comparing List and Endorsement Experiment Data

Function to conduct a statistical test with the null hypothesis that there is no difference between the correlation coefficients between list experiment and endorsement experiment data.

comp.listEndorse(y.endorse, y.list, treat, n.draws = 10000, alpha = 0.05,
  endorse.mean = FALSE, method = "pearson")

Arguments

y.endorse
A numerical matrix containing the response data for the endorsement experiment.

y.list
A numerical vector containing the response data for a list experiment.

treat
A numerical vector containing the binary treatment status for the experiments. The treatment assignment must be the same for both experiments to compare across experiments.

n.draws
Number of Monte Carlo draws.

alpha
Confidence level for the statistical test.

endorse.mean
A logical value indicating whether the mean endorsement experiment response is taken across questions.

method
The method for calculating the correlation, either Pearson's rho or Kendall's tau.

Value

comp.listEndorse returns a list with four elements: the correlation statistic (rho or tau) for the treatment group as cor.treat, the correlation statistic for the control group as cor.control, the p.value for the statistical test comparing the two correlation statistics as p.value, and the bootstrapped confidence interval of the difference as ci.

Details

This function allows the user to calculate the correlation between list and endorsement experiment data within the control group and the treatment group, and to conduct a statistical test with the null hypothesis of no difference between the two correlation coefficients.

References

Blair, Graeme, Jason Lyall and Kosuke Imai. (2014) ``Comparing and Combining List and Experiments: Evidence from Afghanistan." American Journal of Political Science. available at http://imai.princeton.edu/research/comp.html