Item Count Technique

Function to conduct a statistical test with the null hypothesis that there is no "design effect" in a list experiment, a failure of the experiment.

ict.test(y, treat, J = NA, alpha = 0.05, n.draws = 250000, gms = TRUE,
  pi.table = TRUE)

Arguments

y
A numerical vector containing the response data for a list experiment.

treat
A numerical vector containing the binary treatment status for a list experiment.

J
Number of non-sensitive (control) survey items.

alpha
Confidence level for the statistical test.

n.draws
Number of Monte Carlo draws.

gms
A logical value indicating whether the generalized moment selection procedure should be used.

pi.table
A logical value indicating whether a table of estimated proportions of respondent types with standard errors is displayed.

Value

ict.test returns a numerical scalar with the Bonferroni-corrected minimum p-value of the statistical test.

Details

This function allows the user to perform a statistical test on data from a list experiment or item count technique with the null hypothesis of no design effect. A design effect occurs when an individual's response to the non-sensitive items changes depending upon the respondent's treatment status.

References

Blair, Graeme and Kosuke Imai. (2012) ``Statistical Analysis of List Experiments." Political Analysis, Vol. 20, No 1 (Winter). available at http://imai.princeton.edu/research/listP.html

See also

ictreg for list experiment regression based on the assumption of no design effect

Examples

data(affirm)
data(race)

# Conduct test with null hypothesis that there is no design effect
# Replicates results on Blair and Imai (2010) pg. 30

test.value.affirm <- ict.test(affirm$y, affirm$treat, J = 3, gms = TRUE)
print(test.value.affirm)

#> 
#> Test for List Experiment Design Effects
#> 
#> Estimated population proportions 
#>                            est.   s.e.
#> pi(Y_i(0) = 0, Z_i = 1) -0.0019 0.0070
#> pi(Y_i(0) = 1, Z_i = 1)  0.0882 0.0229
#> pi(Y_i(0) = 2, Z_i = 1)  0.1696 0.0287
#> pi(Y_i(0) = 3, Z_i = 1)  0.2388 0.0177
#> pi(Y_i(0) = 0, Z_i = 0)  0.0155 0.0051
#> pi(Y_i(0) = 1, Z_i = 0)  0.1359 0.0155
#> pi(Y_i(0) = 2, Z_i = 0)  0.2073 0.0271
#> pi(Y_i(0) = 3, Z_i = 0)  0.1466 0.0267
#> 
#>  Y_i(0) is the (latent) count of 'yes' responses to the control items. Z_i is the (latent) binary response to the sensitive item.
#> 
#> Bonferroni-corrected p-value
#> If this value is below alpha, you reject the null hypothesis of no design effect. If it is above alpha, you fail to reject the null.
#> 
#> Sensitive Item TRUE 
#>           0.7881128 
#> 

test.value.race <- ict.test(race$y, race$treat, J = 3, gms = TRUE)
print(test.value.race)

#> 
#> Test for List Experiment Design Effects
#> 
#> Estimated population proportions 
#>                            est.   s.e.
#> pi(Y_i(0) = 0, Z_i = 1) -0.0169 0.0084
#> pi(Y_i(0) = 1, Z_i = 1)  0.0101 0.0243
#> pi(Y_i(0) = 2, Z_i = 1)  0.0200 0.0281
#> pi(Y_i(0) = 3, Z_i = 1)  0.0545 0.0091
#> pi(Y_i(0) = 0, Z_i = 0)  0.0304 0.0069
#> pi(Y_i(0) = 1, Z_i = 0)  0.2140 0.0174
#> pi(Y_i(0) = 2, Z_i = 0)  0.3569 0.0263
#> pi(Y_i(0) = 3, Z_i = 0)  0.3309 0.0220
#> 
#>  Y_i(0) is the (latent) count of 'yes' responses to the control items. Z_i is the (latent) binary response to the sensitive item.
#> 
#> Bonferroni-corrected p-value
#> If this value is below alpha, you reject the null hypothesis of no design effect. If it is above alpha, you fail to reject the null.
#> 
#> Sensitive Item 1 
#>       0.04431109 
#>