The 2012 Mexico Elections Panel Study

This dataset is a subset of the 2012 Mexico Elections Panel Study and contains a list experiment question. It reads as follows: I am going to read you a list of four activities that appear on this card and I want you to tell me how many of these activities you have done in recent weeks. Please don''t tell me which ones, just HOW MANY. The four activities are... (SHOW CARD AND READ) a. See television news that mentions a candidate b. Attend a campaign event c. Exchange your vote for a gift, favor, or access to a service d. Talk about politics with other people

Format

A data frame containing the following variables for 1004 observations.

y 	 numeric 	 the number of items that make respondents angry

0 - 4 	 mex.t 	 numeric

treatment status 	 0-1 	 mex.male

numeric 	 whether or not a respondent is male 	 0-1

mex.age 	 numeric 	 age of a respondent

mex.education 	 numeric 	 respondent's level of education

0-9 	 mex.y.all 	 numeric

the number of activities that respondent did 	 0 - 4 	 mex.vote

numeric 	 respondent's self-reported turnout 	 0-1

mex.age2 	 numeric 	 age of a respondent, squared

mex.interest 	 numeric 	 how interested respondent is in politics

1-4 	 mex.married 	 numeric

indicator for whether respondent is married 	 0-1 	 mex.pidpanw2

numeric 	 indicator for whether respondent identifies with PAN party 	 0-1

mex.pidprdw2 	 numeric 	 indicator for whether respondent identifies with PRD party

0-1 	 mex.pidpriw2 	 numeric

indicator for whether respondent identifies with PRI party 	 0-1 	 mex.votecard

numeric 	 respondent's enumerator-verified turnout 	 0-1

mex.urban 	 numeric 	 indicator for whether respondent lives in urban area

0-1 	 mex.cleanelections 	 numeric

indicator for whether respondent thinks elections were clean 	 0-1 	 mex.cleanelectionsmiss

numeric 	 indicator for whether cleanelections variable was missing 	 0-1

mex.metro 	 numeric 	 indicator for whether respondent lives in Mexico City metro area

0-1 	 mex.centralregion 	 numeric

indicator for whether respondent lives in Mexico's central region 	 0-1 	 mex.northregion

numeric 	 indicator for whether respondent lives in Mexico's north region 	 0-1

mex.wealth 	 numeric 	 scale for respondent's wealth, based on household asset indicators

mex.epnapprove 	 numeric 	 respondent's approval rating of Enrique Pena-Nieto

1-11 	 mex.havepropoganda 	 numeric

indicator for whether respondent has propaganda outside their home 	 0-1 	 mex.concurrent

numeric 	 indicator for whether respondent lives in state with concurrent elections 	 0-1

mex.loyal 	 numeric 	 indicator for whether respondent strongly identifies with the PAN, PRI, or PRD party

0-1 	 y 	 numeric

Source

The full data set is available at the Mexico Panel Study website (http://mexicopanelstudy.mit.edu/)

Details

where item c. is presented only to the treatment group, and the control list only contains the other three items.