The 1994 Multi-Investigator Survey

This dataset is a subset of the 1994 Multi-Investigator Survey and contains the item count technique or the list experiment. The main question reads as follows: Now I'm going to read you four things that sometimes make people angry or upset. After I read all (three/four), just tell me HOW MANY of them upset you. (I don't want to know which ones, just how many.) (1) "the federal government increasing the tax on gasoline;" (2) "professional athletes getting million-dollar-plus salaries;" (3) "requiring seatbelts be used when driving;" (4) "large corporations polluting the environment;" (5) "black leaders asking the government for affirmative action."

Format

A data frame containing the following 6 variables for 1171 observations.

y 	 numeric 	 the number of items that make respondents angry

0 - 4 	 sensitive 	 numeric

whether or not the sensitive item (asked directly) makes respondents angry 	 0 - 1 	 south

numeric 	 whether or not a respondents live in a southern state 	 0 - 1

male 	 numeric 	 whether or not a respondent is male

0 - 1 	 college 	 numeric

whether or not a respondent attended some college 	 0 - 1 	 age

numeric 	 age of a respondent divided by 10 	 democrat

numeric 	 whether not a respondent identifies as a Democrat 	 0 - 1

republican 	 numeric 	 whether not a respondent identifies as a Republican

0 - 1 	 independent 	 numeric

whether not a respondent identifies as an independent 	 0 - 1 	 treat

numeric 	 treatment status 	 0 - 1

list.data 	 numeric 	 indicator for list experiment subset (treatment and control groups)

0 - 1 	 sens.data 	 numeric

indicator for direct sensitive item subset 	 0 - 1 	 y

Source

The full data set is available at SDA (Survey Documentation and Analysis; http://sda.berkeley.edu/D3/Multi/Doc/mult.htm)

Details

where the last item is presented only with the treatment group and the control list only contains the first three items.

The survey also includes a question in which attitudes toward the sensitive item are asked directly. Now I'm going to ask you about another thing that sometimes makes people angry or upset. Do you get angry or upset when black leaders ask the government for affirmative action?